api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic>=0.40.0
python-pptx>=0.6.23
pydantic>=2.0.0
python-multipart>=0.0.9
